requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
replicate==0.22.0
python-dotenv==1.0.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
pydantic==2.5.0
